% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_C_index.R
\name{get_C_index}
\alias{get_C_index}
\title{get_C_index: Calculate Harrel's C-index.}
\usage{
get_C_index(cov, beta, time, delta)
}
\arguments{
\item{cov}{The covariate matrix.}

\item{beta}{The estimated beta vector.}

\item{time}{Time vector}

\item{delta}{a vector of the censoring indicator corresponding to time points of dH. 0 indicating censoring, while 1 indicating non-censoring.}
}
\description{
Calculate the Harrel's concordance index(C-index) based on the estimated beta
}
\examples{
n <- 100
time <- rexp(n, rate = 0.1)  # Survival times
status <- sample(0:1, n, replace = TRUE)  # Censoring indicator
X <- matrix(rnorm(n * 2), n, 2)  # Two covariates
cox_model <- coxph(Surv(time, status) ~ X[,1] + X[,2])
beta <- summary(cox_model)$coefficients[, 1]
C <- get_C_index(X, beta, time, status)
}
\author{
Yuxi Song. Email: lucillesong@ucla.edu
}
