% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_complete_loss.R
\name{get_complete_loss}
\alias{get_complete_loss}
\title{get_complete_loss: Calculate loss function using negative complete log likelihood.}
\usage{
get_complete_loss(delta, cov, beta, haz)
}
\arguments{
\item{delta}{a vector of the censoring indicator corresponding to time points of dH. 0 indicating censoring, while 1 indicating non-censoring.}

\item{cov}{The covariate matrix.}

\item{beta}{The estimated beta vector.}

\item{haz}{The estimated hazard vector at each time point}
}
\description{
Calculate the cumulative hazard function given the hazard and the censoring indicator at each time point. If the censoring indicator is not given, calculate the cumulative hazard as the sum of the hazard at time points prior to the time.
}
\examples{
h <- c(0.2, 0.1, 0.3)
delta <- c(0, 1, 0)
dat <- dH_to_cumH(dH = h, delta = delta)

}
\author{
Yuxi Song. Email: lucillesong@ucla.edu
}
